% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_captions}
\alias{add_captions}
\title{Add captions (title, notes, etc) to a plot or grid graphics object (grob)}
\usage{
add_captions(plot, theme, title = NULL, subtitle = NULL, note = NULL,
  source = NULL, source_url = NULL, show.logo = TRUE)
}
\arguments{
\item{plot}{the \code{ggplot} plot object or a {grid} grob}

\item{theme}{the \code{ggplot} theme object to use for styling captions}

\item{title}{the plot title (displayed at top)}

\item{subtitle}{the plot subtitle}

\item{note}{the plot note}

\item{source}{the plot source}

\item{source_url}{the plot source URL or other location info}

\item{show.logo}{if TRUE, include a logo}
}
\value{
a new \code{gtable} object which can be \code{print}ed to a graphic device
}
\description{
\code{add_captions} adds captions to a plot outside of \code{ggplot}. This is
useful since the default \code{ggplot} title, subtitle and caption align with
the plot area, rather than the entire figure (including axes),
unlike many publishing standards.
}
\details{
The \code{title} and \code{subtitle} arguments correspond directly to the
same in \code{title}, and appear above the plot. The \code{note} and
\code{source} arguments share the work of \code{ggplot}'s \code{caption} and
appear below the plot. The \code{source_url} and logo appear at the very
bottom of the plot, in slightly smaller text, and would usually only be
shown in standalone, branded graphics (for e.g. distribution on social media)
and not, for instance, within a report.

The relevant theme elements within \code{theme} are the same as would be used
if the the \code{ggplot} title, subtitle and caption were used.
}
\examples{
library(ggplot2)
library(grid)

# Simple example for adapting a ggplot labs() call...
p <- ggplot(mtcars, aes(cyl, mpg)) + geom_point()

p + labs(
  title = "This is a title",
  subtitle = "This is a subtitle",
  caption = "This is a caption"
)

# ...to use add_captions() instead
g <- add_captions(
  plot = p,
  theme = p$theme,
  title = "This is a title",
  subtitle = "This is a subtitle",
  note = "Note: This is a note",
  source = "Note: This is a source",
  source_url = "http://where.this.is.found.com",
  show.logo = TRUE
)
grid.newpage()
grid.draw(g)

# Some relevant text attributes of ggplot themes are obeyed
p <- ggplot(mtcars, aes(cyl, mpg)) + geom_point() +
  theme_classic() + theme(plot.title = element_text(color = "blue"))

p + labs(
  title = "This is a title",
  subtitle = "This is a subtitle",
  caption = "This is a caption"
)

g <- add_captions(
  plot = p,
  theme = p$theme,
  title = "This is a title",
  subtitle = "This is a subtitle",
  note = "Note: This is a note",
  source = "Note: This is a source",
  show.logo = FALSE
)
grid.newpage()
grid.draw(g)
}
