% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{bracketed_breaks}
\alias{bracketed_breaks}
\title{A breaks function for ggplot2 scales that shows the data limits}
\usage{
bracketed_breaks(limits, at = 1, closeness_threshold = 0.1,
  breaker = scales::extended_breaks())
}
\arguments{
\item{limits}{A pair of limits, overrides the default}

\item{at}{To limit the breaks to be on 5s, 10s, etc (more of a hint)}

\item{closeness_threshold}{If one of the data limit breaks and one of the
default breaks will be closer than closeness_threshold * the ggplot2
calculated limits, the the nearby default break will be dropped.}
}
\description{
This builds on the standard ggplot2 continuous breaks (extended breaks)
function to make sure the data limits, given as a parameter, are included
in the breaks.
}
\details{
By default it will try to use the limits ggplot provides, but by default those
are extended by the expand argument of scales, so it will use the integral
ceiling and floor respectively of these. Usually this will work. If not you
can provide limits to override.
}
