% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{comb_aes}
\alias{comb_aes}
\alias{comb_aes_values}
\title{Combine factors for a single \code{ggplot} aesthetic}
\usage{
comb_aes(...)

comb_aes_values(..., FUN, not.found = NA)
}
\description{
Combine factors for a single \code{ggplot} aesthetic
}
\examples{
# Suppose we want a fill that uses two variables
p <- ggplot(
  diamonds[1:10,],
  aes(depth, price, color = comb_aes(cut, color))
) +
  geom_point()
p

# If we have labels set up for each of the factors, we can combine them
# (although this doesn't offer much advantage over using a function for
# the label argument).

cut.labels <- setNames(as.character(unique(diamonds$cut)), unique(diamonds$cut))
color.labels <- c(
  D = "D (best)", E = "E", F = "F", G = "G", H = "H", I = "I", J = "J (best)"
)
labels <- comb_aes_values(diamonds$cut, diamonds$color, FUN =
  function(cut, color) {
    paste0(cut.labels[[cut]], " / ",color.labels[[color]])
  }
)
p + scale_color_discrete(labels = labels)

# More usefully, we can do the same with other aesthetics (e.g. color) to
# create multi-dimensional color scales.

}
