% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contrasting_colors}
\alias{contrasting_colors}
\title{Select contrasting text colors, from a list, for a set of fill colors}
\usage{
contrasting_colors(fillcolors, textcolors = c("black", "white"))
}
\description{
Given a (possibly named) vector of fill (background) colors, this function
returns a (possibly named) vector of highest-contrast text (foreground)
colors, from a list of potential text colors (usually two is sufficient).
}
