% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_lined_area}
\alias{geom_lined_area}
\title{TODO: option to only line adjacent areas with same fill}
\usage{
geom_lined_area(color = "white", size = 0.35, ...)
}
\description{
TODO: option to only line adjacent areas with same fill
}
