% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_other_dotplot}
\alias{geom_other_dotplot}
\title{A geom for the other kind of dotplot, with a connecting line}
\usage{
geom_other_dotplot(mapping = aes(), linecolor = "black", linesize = 0.25,
  arrow = NA, flip.legend = FALSE, size = 1, ...)
}
\arguments{
\item{mapping}{Complicatedly, you must specify x (the value axis), y (the
categorical axis) and group in the geom's own mapping argument, even if
you specify them in the main ggplot() call. If the chart is faceted, you
need to also include the facetting variable as part of the group (use paste0).}

\item{linecolor}{The color of the grouping line}

\item{linesize}{The size of the grouping line}

\item{arrow}{NA or FALSE for no arrow, TRUE for a standard arrow, or pass
a grob for something weird and wonderful. For reverse direction groups, the}

\item{flip.legend}{By default the legend will show the arrow facing forward,
but if most the groups run backwards it may be helpful to show the arrow
backwards too, so you can set this to TRUE to do that.}

\item{size}{The size of the points & arrows if used.}
}
\description{
This is still a fairly hacky way to make dotplots, so you have to observe
and obey some weirdness, which is outlined below. Also, the "direction"
(usually time) of the data is dictated by the row ordering in the data frame,
so make sure it is sorted correctly.
}
