% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{repel}
\alias{repel}
\title{Repel the values in a vecto away from each other}
\usage{
repel(x, gap = 1, force = FALSE, err = 0.01)
}
\description{
Moves the values in a vector so that they each have gap space between them.
If force is FALSE (default), it will warn and do nothing if this can't be
done within a space of range(x) + c(-gap, gap).
}
