% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rgba2rgb}
\alias{rgba2rgb}
\title{Convert an RGBA color to RGB by combining with a background color}
\usage{
rgba2rgb(color, background = "white", alpha = NULL, preserve.names = TRUE)
}
\arguments{
\item{color}{a color or vector of colors}

\item{background}{a background color}

\item{alpha}{the alpha to use, which if not NULL will override any alphas in \code{color}}

\item{preserve.names}{preserve the names of \code{color} if any}
}
\description{
Convert an RGBA color to RGB by combining with a background color
}
