% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{str_wrap_lines}
\alias{str_wrap_lines}
\title{Wrap a vector of strings to a specified number of lines}
\usage{
str_wrap_lines(strings, lines = 2, indent = 0, exdent = 0,
  force = FALSE)
}
\arguments{
\item{force}{Force strings to take up \code{lines} lines, even if they could
fit the width with fewer lines}
}
\description{
Like \code{stringr::str_wrap()} except that instead of specifying a width you
specify a "height" (ie. a number of lines that each string can occupy). The
longest string and "height" will jointly determine the width, which will then
be applied to the other strings - so that strings are not unnecessarily wrapped.
}
