% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apihelper.R
\name{wbg_source}
\alias{wbg_source}
\title{Build a source string for World Bank indicators}
\usage{
wbg_source(indicatorIDs, source = NULL)
}
\arguments{
\item{indicatorIDs}{a vector of indicator IDs (SETS codes)}
}
\description{
\code{wbg_source} builds a source string for a set of indicators.
}
\details{
For each indicator, it takes the source from the API metadata, then extracts the
part leading up to the first period (.), which is usually useable as a short
source. These are then combined into a single string

Since population-related indicators are commonly included as references, this
function excludes them from source string construction.
}
