% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgcharts.R
\docType{package}
\name{wbgcharts-package}
\alias{wbgcharts}
\alias{wbgcharts-package}
\title{wbgcharts: Chart styles & enhancements for the World Bank Group}
\description{
Themes, extra geoms and post-production finishing tools for use with ggplot2
}
\author{
\strong{Maintainer}: Andrew Whitby \email{awhitby@worldbank.org}

}
\keyword{internal}
