% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apihelper.R
\name{wbgdata}
\alias{wbgdata}
\title{Download data from the World Bank API (wrapper for wbstats::wb)}
\usage{
wbgdata(country = "all", indicator, startdate, enddate, years, ...,
  col.indicator = FALSE, cache = get_wbcache(), indicator.wide = TRUE,
  removeNA = FALSE, offline = "none", offline.file,
  rename.indicators = FALSE)
}
\arguments{
\item{country}{Character vector of country or region codes. Default value is special code of \code{all}.
Other permissible values are codes in the following fields from the \code{\link{wb_cachelist}} \code{country}
data frame. \code{iso3c}, \code{iso2c}, \code{regionID}, \code{adminID}, and \code{incomeID}.
Additional special values include \code{aggregates}, which returns only aggregates, and \code{countries_only},
which returns all countries without aggregates.}

\item{indicator}{a vector of indicator (SETS) codes. If named and
\code{rename.indicators} is true, the indicator codes will be replaced
by their respective names.}

\item{startdate}{Numeric or character. If numeric it must be in \%Y form (i.e. four digit year).
For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01"
and for quarterly data, "2016Q1". This also accepts a special value of "YTD", useful for more frequently
updated subannual indicators.}

\item{enddate}{Numeric or character. If numeric it must be in \%Y form (i.e. four digit year).
For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01"
and for quarterly data, "2016Q1".}

\item{years}{an alternative to \code{startdate} and \code{enddate} (which it
overrides), a vector of years for which to download data.}

\item{col.indicator}{if true, return the indicator (description) column}

\item{cache}{the wbstats cache object to use (default usually ok)}

\item{indicator.wide}{return the indicators in wide (not long) format}

\item{removeNA}{if \code{TRUE}, remove any blank or \code{NA} observations that are returned.
if \code{FALSE}, no blank or \code{NA} values are removed from the return.}

\item{offline}{(experimental) can be "none", "only", "warn" or "overwrite"}

\item{rename.indicators}{see \code{indicator} paramater}
}
\description{
\code{wbgdata} is a wrapper for the \code{wbstats::wb} function. It also
downloads data using the World Bank API, but offers enhancements including
better cached data management, more input parameter formats and more output
formats, based boilerplate code I found myself constantly retypting.
}
